requirements.txt
Pillow
opencv-python
natsort